Refactoring code into multiple files
