+++ commands/utility/ping.js
@@ -1,0 +1,7 @@
+module.exports = {
+	name: 'ping',
+	description: 'Ping!',
+	execute(message, args) {
+		message.channel.send('Pong.');
+	},
+};
